fix R and Latex compilers
